Develop (#23)
* Убраны лишние данные из регистрации пользователя: оставлен только email, телефон и пароль

* Удален WeatherForecastController

* UserService

* Обновление

* CRUD операции по пользователям

* Обновка миграций

* Удаление лишних ссылок из зоны ответственности пользователей, миграции

* Контроллер машин + обновление контроллера пользователей

* Удаление лишнего из метода удаления пользователя :)

* Приведение моделей и логики в соответствие с реалиями :) Протестированы все методы

* Apartment service dronesaurum (#14)

* ApartmentService

* Контроллер сервиса квартир

* Обновление контроллера, добавление поддержки null для листа пользаков в модели, обновление DBcontext

* Добавление моделей, добавление поддержки RabbitMQ

* Почти готово, допиливаю методы, которые связываются с сервисом пользователей

* Допилены DTO, десериализация JSON, а также метод получения квартир пользователя по его телефону. Ну бомба просто

* Обновление сигнатуры метода добавления пользователя к квартире

* Add Gateway & Notification Service

* Add Notification Service Base Implement

* Add RabbitMQ to Notification Service

* Update RabbitMQ Consumer

Сделал его генериком

* Update InMemoryNotificationRepository.cs

* Update .gitignore

* Update .gitignore and .env

* Update Notification Service

* Update appsettings

* Indentity: Register and Login pages (#17)

* Обновление Id пользователя на Guid, добавление пользователя в UserService при создании в Identity

* Страницы регистрации и логина, компоненты JS/CSS/Head, стилизация

* Регистрация через форму работает!!1

* Рефакторинг IAccountService

* Редирект на основную страницу сервиса, который будет фронтом после авторизации

* Гитигнор почему-то потерялся

* Front main page (#19)

* Добавление env-файла

* Удаление дубля в compose

* Удоление лишнего из AuthController

* Надо разобраться с редиректом до логина и проверить работоспособность

* Авторизация проходится

* Обновка областей видимости во вьюхе

* Merge front and develop (#22)

* Добавление env-файла

* Удаление дубля в compose

* Удоление лишнего из AuthController

* Надо разобраться с редиректом до логина и проверить работоспособность

* Авторизация проходится

* Обновка областей видимости во вьюхе

* Контроллер с маршрутизацией и страницы фронта

* Настройка вебпака

* Тяну данные пользака с микрсоервиса пользаков на фронт!

* Переписанные скрипты, но осталось понять почему дубли в вызове идут

* Реализовано нормальное подтягивание данных пользователя в общий профиль на фронте

* Начало работы над страницей редактирования пользователя

* Верстка профиля пользователя

* Обновление модельки пользователя, наполнение профиля пользователя, коллекционирование данных перед сохранением

* Обновление модели пользователя + добавление класса валидации

* Добавление машин сделано, осталась валидация и удаление

* Удоление и уведомления о статусе сделаны, осталась валидация данных

* Изменение названий JS

* Допилена валидация

* Какие-то лищние файлы

* Мерж фронта и девелопа

---------

+++ HPM-System.UserService/Controllers/CarsController.cs
@@ -74,7 +74,7 @@ namespace HPM_System.UserService.Controllers
 
         // GET: api/Cars/{id}
         [HttpGet("{id}")]
-        public async Task<IActionResult> GetCar(int id)
+        public async Task<IActionResult> GetCar(long id)
         {
             try
             {
@@ -146,7 +146,7 @@ namespace HPM_System.UserService.Controllers
 
         // DELETE: api/Cars/{id}
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteCar(int id)
+        public async Task<IActionResult> DeleteCar(long id)
         {
             try
             {
